backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
debugpy==1.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0

# LangChain and RAG
langchain==0.0.350

# Vector Database (PostgreSQL with pgvector)
# qdrant-client==1.7.0  # Removed - using PostgreSQL with pgvector

# Document Processing
pypdf==3.17.4
markdown==3.5.1
python-magic==0.4.27

# PDF Processing with OCR
pdf2image==1.16.3
pix2text==0.2.0
Pillow==10.0.1
opencv-python-headless==4.8.1.78

# Embeddings - Use sentence-transformers for better embeddings
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
numpy==1.23.5
huggingface-hub==0.16.4
# pgvector-python for better PostgreSQL integration
pgvector==0.2.4

# Notion Integration
notion-client==2.2.1

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9 